update comment code

+++ SE1614_Group4_Project_FE/tech-blog/js/comment.js
@@ -2,6 +2,7 @@ var userId, postId;
 $(document).ready(() => {
     var token = localStorage.getItem("token");
 
+    if (token != null) {
     let decoded = jwt_decode(token);
     let userRole =
         decoded[
@@ -11,6 +12,8 @@ $(document).ready(() => {
     userId = decoded[
         "http://schemas.xmlsoap.org/ws/2005/05/identity/claims/sid"
     ];
+    }
+
     let urlParam = new URLSearchParams(window.location.search);
     postId = urlParam.get("id");
 
@@ -19,41 +22,57 @@ $(document).ready(() => {
 })
 
 var commentDto;
-function AddComment() {
+
+function AddComment(model) {
+   var userReplyId = model.getAttribute("data-id-reply");
         var content = $("#contentComment").val();
         commentDto = {
             content: content,
             postId: postId,
             userId: userId,
+            replyUserId: userReplyId
         }
         if (content != "") {
-            $.ajax({
-                url: "https://localhost:7065/api/Comment/AddComment",
-                method: "POST",
-                contentType: "application/json",
-                headers: {
-                    "Authorization": "Bearer " + localStorage.getItem('token')
-                },
-                data: JSON.stringify(commentDto),
-                success: function (response) {
-                    SlimNotifierJs.notification(
-                        "success",
-                        "Comment Successfully!",
-                        response,
-                        3000
-                    );
-                    $("#contentComment").val("");
-                    loadDataComment();
-                },
-                error: function (xhr, status, error) {
-                    SlimNotifierJs.notification(
-                        "error",
-                        "Error",
-                        xhr.responseText,
-                        3000
-                    );
-                },
-            });
+            if (token !== "") {
+                $.ajax({
+                    url: "https://localhost:7065/api/Comment/AddComment",
+                    method: "POST",
+                    contentType: "application/json",
+                    headers: {
+                        "Authorization": "Bearer " + localStorage.getItem('token')
+                    },
+                    data: JSON.stringify(commentDto),
+                    success: function (response) {
+                        SlimNotifierJs.notification(
+                            "success",
+                            "Comment Successfully!",
+                            response,
+                            3000
+                        );
+                        $("#contentComment").val("");
+                        loadDataComment();
+                        refesh();
+                        $("#replyComment").empty();
+
+                    },
+                    error: function (xhr, status, error) {
+                        SlimNotifierJs.notification(
+                            "error",
+                            "Error",
+                            xhr.responseText,
+                            3000
+                        );
+                    },
+                });
+            } else {
+                SlimNotifierJs.notification(
+                    "error",
+                    "Error",
+                    "Please login!!",
+                    3000
+                );
+            }
+          
         } else {
             SlimNotifierJs.notification(
                 "error",
@@ -64,7 +83,48 @@ function AddComment() {
         }
     
 }
+function replyComment(model) {
+    var userReplyId = model.getAttribute("data-user-reply-id");
 
+    if (token !== "") {
+        $.ajax({
+            url: "https://localhost:7065/api/Comment/getUserReplyComment/" + userReplyId,
+            method: "GET",
+            contentType: "application/json",
+            headers: {
+                "Authorization": "Bearer " + localStorage.getItem('token')
+            },
+            success: function (response) {
+                $("#replyComment").empty();
+                $("#replyComment").append(`<p> Reply comment:   ${response.username} </p>` );
+
+                $("#btnUpdate").empty();
+                $("#btnUpdate").append(
+                    `<button class="btn btn-primary" data-id-reply=${userReplyId} onclick="AddComment(this)"  type="button">Comment</button>
+                    <button style="margin-left: 5px;" class="btn btn-delete" onclick="refesh()"  type="button">Cancel</button>
+                    `
+                )
+
+
+            },
+            error: function (xhr, status, error) {
+                SlimNotifierJs.notification(
+                    "error",
+                    "Error",
+                    xhr.responseText,
+                    3000
+                );
+            },
+        });
+    } else {
+        SlimNotifierJs.notification(
+            "error",
+            "Error",
+            "Please login!!",
+            3000
+        );
+    }
+}
 
 var page = 1;
 var pageSize = 3;
@@ -101,18 +161,23 @@ function loadDataComment() {
                         >
                     </h4>
                     <p>
+                    
+                    <b> ${item?.userNameReply != null ?  item?.userNameReply+" "  : ""}
+                    </b>
                     ${item?.content}
                     </p>
                     <div id="btnCommentUpdate">
-                    <a
-                    href="#"
+                    <button
+                    data-user-reply-id=${item.userId}
+                    onclick="replyComment(this)"
                     class="btn btn-primary btn-sm"
-                    >Reply</a>
+                    >Reply</button>
                     ${ item?.userId === userId ? 
                         `<button
                         class="btn btn-primary btn-sm"
                         data-comment-id=${item.id}
                         data-comment-content="${item?.content}"
+                        data-reply=${item.replyUserId}
                         onclick="updateComment(this)"
                         >Update</button>
                         
@@ -182,56 +247,64 @@ function loadDataComment() {
 
 function deleteComment(model) {
    var commentId = model.getAttribute("data-comment-id");
-   $.ajax({
-    url: "https://localhost:7065/api/Comment/DeleteComment/" + commentId, 
-    method: "DELETE",
-    contentType: "application/json",
-    headers: {
-        "Authorization": "Bearer " + localStorage.getItem('token')
-    },
-    success: (response) => {
-        SlimNotifierJs.notification(
-            "success",
-            "Success",
-            "Delete success",
-            3000
-        );
-        loadDataComment();
-
-
-    },
-    error: function (xhr, status, error) {
-        SlimNotifierJs.notification(
-            "error",
-            "Error",
-            "Failed to delete",
-            3000
-        );
-    },
-});
+   if (confirm("Delete this comment??") == true) {
+    $.ajax({
+        url: "https://localhost:7065/api/Comment/DeleteComment/" + commentId, 
+        method: "DELETE",
+        contentType: "application/json",
+        headers: {
+            "Authorization": "Bearer " + localStorage.getItem('token')
+        },
+        success: (response) => {
+            SlimNotifierJs.notification(
+                "success",
+                "Success",
+                "Delete success",
+                3000
+            );
+            loadDataComment();
+    
+    
+        },
+        error: function (xhr, status, error) {
+            SlimNotifierJs.notification(
+                "error",
+                "Error",
+                "Failed to delete",
+                3000
+            );
+        },
+    });
+   } else {
+  }
+  
 }
 var commentId;
 function updateComment(model) {
     commentId = model.getAttribute("data-comment-id");
+    var replyUserId = model.getAttribute("data-reply");
     var contentUpdate = model.getAttribute("data-comment-content");
-    console.log(contentUpdate)
     $("#contentComment").val(contentUpdate);
 
     $("#btnUpdate").empty();
     $("#btnUpdate").append(
-        `<button class="btn btn-primary" onclick="update()"  type="button">Comment</button>
+        `<button class="btn btn-primary" data-id=${replyUserId} onclick="update(this)"  type="button">Comment</button>
         <button style="margin-left: 5px;" class="btn btn-delete" onclick="refesh()"  type="button">Cancel</button>
         `
     )
 
  }
 
- function update() {
+ 
+ function update(model) {
+    var replyUserId = model.getAttribute("data-id");
     var updateComment = {
         userId: userId,
         postId: postId,
         content: document.getElementById("contentComment").value,
+        replyUserId: replyUserId
     }
+    console.log(updateComment.replyUserId)
     $.ajax({
         url: "https://localhost:7065/api/Comment/UpdateComment/" + commentId, 
         method: "PUT",
@@ -266,7 +339,9 @@ function updateComment(model) {
     $("#contentComment").val("");
     $("#btnUpdate").empty();
     $("#btnUpdate").append(
-        `<button class="btn btn-primary" onclick="AddComment()" id="btnComment" type="button">Comment</button>
+        `<button class="btn btn-primary" onclick="AddComment(this)" id="btnComment" type="button">Comment</button>
         `
     )
+    $("#replyComment").empty();
+
  }
